Move ftp config properties
