+++ src/test/java/uk/gov/hmcts/reform/sendletter/services/FtpClientTest.java
@@ -10,10 +10,10 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
+import uk.gov.hmcts.reform.sendletter.config.FtpConfigProperties;
 import uk.gov.hmcts.reform.sendletter.exception.FtpException;
 import uk.gov.hmcts.reform.sendletter.model.Report;
 import uk.gov.hmcts.reform.sendletter.services.zip.ZippedDoc;
-import uk.gov.hmcts.reform.slc.config.FtpConfigProperties;
 
 import java.io.IOException;
 import java.util.List;
